+++ src/componentes/escolas/Despesas/CadastroDeDespesasPipeline/index.js
@@ -1,13 +1,9 @@
 import React from "react";
-import {CadastroForm} from "./CadastroForm";
+import {DespesaFormPipeline} from "./DespesaFormPipeline";
 
-export const CadastroDeDespesasPipeline = ({verbo_http, veioDeSituacaoPatrimonial}) =>{
-
-    return (
-        <CadastroForm
-            verbo_http={verbo_http}
-            veioDeSituacaoPatrimonial={veioDeSituacaoPatrimonial}
-        />
-    );
-
-}
+export const CadastroDeDespesasPipeline = ({verbo_http, veioDeSituacaoPatrimonial}) => (
+    <DespesaFormPipeline
+        verbo_http={verbo_http}
+        veioDeSituacaoPatrimonial={veioDeSituacaoPatrimonial}
+    />
+);
