+++ src/componentes/escolas/Despesas/CadastroDeDespesasPipeline/modals/ModalDespesaJaConferida.js
@@ -1,5 +1,5 @@
 import React from "react";
-import {ModalBootstrap} from "../../../Globais/ModalBootstrap";
+import {ModalBootstrap} from "../../../../Globais/ModalBootstrap";
 
 export const ModalDespesaConferida = (props) =>{
     return (
