+++ src/componentes/escolas/Despesas/CadastroDeDespesasPipeline/utils/index.js
@@ -1,0 +1,42 @@
+/**
+ * Helpers do CadastroDeDespesasPipeline.
+ * Preferir imports nomeados daqui em vez de arquivos soltos em src/utils.
+ */
+
+export {
+    convertToNumber,
+    round,
+    trataNumericos,
+    calculaValorRateio,
+    calculaValorRecursoAcoes,
+    calculaValorOriginal,
+    apenasNumero,
+} from "./numericos";
+
+export {
+    cpfMaskContitional,
+    processoIncorporacaoMask,
+} from "./mascaras";
+
+export {
+    valida_cpf_cnpj,
+    valida_cpf_cnpj_permitindo_cnpj_zerado,
+    valida_cnpj,
+    valida_cpf_exportado,
+} from "./cpfCnpj";
+
+export {comparaObjetos} from "./comparaObjetos";
+
+export {
+    periodoFechado,
+    periodoFechadoImposto,
+} from "./periodoFechado";
+
+export {validaPayloadDespesas} from "./payloadDespesas";
+
+export {
+    YupSignupSchemaCadastroDespesa,
+    YupSignupSchemaCadastroDespesaSaida,
+} from "./schemasDespesa";
+
+export {metodosAuxiliares} from "./metodosAuxiliares";
