--- src/utils/ValidacoesAdicionaisFormulariosDespesasPipeline.js
@@ -1,832 +1,0 @@
-/* eslint eqeqeq: 0 */
-/* eslint-disable */
-import * as yup from "yup";
-import moment from "moment";
-import {ASSOCIACAO_UUID} from "../services/auth.service";
-import {getPeriodoFechado} from "../services/escolas/Associacao.service";
-import {metodosAuxiliares} from "../componentes/escolas/Despesas/metodosAuxiliares";
-
-export const YupSignupSchemaAssociacoes  = yup.object().shape({
-  nome: yup.string().required("Nome é obrigatório"),
-  codigo_eol_unidade: yup.string().required("Código EOL da unidade é obrigatório"),
-  periodo_inicial: yup.string().test('test-name', 'Período inicial é obrigatório',
-      function (value) {
-        return !(value === undefined || value === null || value === "");
-
-      }),
-  cnpj: yup.string()
-  .test('test-name', 'Digite um CNPJ Válido',
-      function (value) {
-        if (value){
-          return valida_cnpj(value)
-        }else {
-          return true
-        }
-      }),
-});
-
-export const YupSignupSchemaDreDadosDiretoria = yup.object().shape({
-  dre_cnpj: yup.string()
-  .test('test-name', 'Digite um CNPJ Válido',
-      function (value) {
-        if (value){
-          return valida_cnpj(value)
-        }
-        else{
-          // O campo CNPJ não é mais obrigatório
-          return true;
-        }
-      }),
-});
-
-export const YupSignupSchemaDadosDaAssociacao = yup.object().shape({
-  email: yup.string().email("Digite um email válido").nullable(),
-});
-
-export const YupSignupSchemaAlterarSenha = yup.object().shape({
-  senha_atual: yup.string().required("Campo senha atual é obrigatório"),
-  senha: yup.string().required("Campo nova senha é obrigatório"),
-  confirmacao_senha: yup.string().required("Campo confirmação da nova senha é obrigatório")
-  .oneOf([yup.ref('senha'), null], 'As senhas precisam ser iguais'),
-});
-
-export const YupSignupSchemaAlterarEmail = yup.object().shape({
-  email: yup.string().required("Campo email é obrigatório"),
-  confirmacao_email: yup.string().required("Campo confirmação do email é obrigatório")
-  .oneOf([yup.ref('email'), null], 'Os emails precisam ser iguais'),
-});
-
-export const YupSignupSchemaRedefinirSenha = yup.object().shape({
-  senha: yup.string().required("Campo nova senha é obrigatório"),
-  confirmacao_senha: yup.string().required("Campo confirmação da nova senha é obrigatório")
-  .oneOf([yup.ref('senha'), null], 'As Senhas precisam ser iguais'),
-});
-
-export const YupSignupSchemaRecuperarSenha = yup.object().shape({
-  usuario: yup.string().required("Campo usuário é obrigatório"),
-});
-
-export const YupSignupSchemaLogin = yup.object().shape({
-  login: yup.string().required("Campo Usuário é obrigatório"),
-  senha: yup.string().required("Campo Senha é obrigatório"),
-});
-
-export const YupSignupSchemaMembros = yup.object().shape({
-  representacao: yup.string().required("Representação é obrigatório"),
-
-  codigo_identificacao: yup.string()
-    .test('test-name', 'É obrigatório e não pode ultrapassar 10 caracteres',
-        function (value) {
-          const { representacao } = this.parent;
-          if(representacao === "SERVIDOR" || representacao === "ESTUDANTE"){
-            return !(!value || value.trim() === "" || value.length > 10);
-          }else {
-            return true
-          }
-      }),
-
-  nome: yup.string()
-  .test('test-name', 'É obrigatório e não pode ultrapassar 160 caracteres',
-      function (value) {
-        const { representacao } = this.parent;
-        if(representacao === "PAI_RESPONSAVEL"){
-          return !(!value || value.trim() === "" || value.length > 160);
-        }else {
-          return true
-        }
-      }),
-
-  cargo_educacao: yup.string()
-  .test('test-name', 'É obrigatório e não pode ultrapassar 45 caracteres',
-      function (value) {
-        const { representacao } = this.parent;
-        if(representacao === "SERVIDOR"){
-            return !(!value || value.trim() === "" || value.length > 45);
-        }else {
-          return true
-        }
-      }),
-});
-
-export const YupSignupSchemaCadastroDespesa = yup.object().shape({
-  cpf_cnpj_fornecedor: yup.string().required("Campo CNPJ ou CPF é obrigatório")
-  .test('test-name', 'Digite um CPF ou um CNPJ válido',
-      function (value) {
-        if(value !== undefined){
-          return valida_cpf_cnpj(value)
-        }else {
-          return true
-        }
-      }),
-  nome_fornecedor: yup.string().nullable(),
-  tipo_documento:yup.string().nullable(),
-  numero_documento:yup.string().nullable(),
-  data_documento: yup.string().nullable(),
-  tipo_transacao: yup.string().nullable(),
-  data_transacao: yup.string().nullable(),
-  documento_transacao: yup.string().nullable(),
-  valor_total: yup.string().nullable(),
-  valor_recursos_proprios: yup.string().nullable(),
-  valor_total_dos_rateios:yup.string().nullable(),
-  valor_recusos_acoes:yup.string().nullable(),
-});
-
-export const YupSignupSchemaCadastroDespesaSaida = yup.object().shape({
-
-  nome_fornecedor: yup.string().required("Nome Fornecedor é obrigatório")
-  .test('test-nome-fornecedor', 'Digite um nome de fornecedor válido',
-      function (value) {
-        if (value !== undefined || value !== '') {
-          return true
-        } else {
-          return false
-        }
-      }),
-
-  tipo_documento:yup.string().required("Tipo de documento é obrigatório")
-  .test('test-tipo-documento', 'Selecione um tipo de documento válido',
-      function (value) {
-        if (value !== undefined || value !== '') {
-          return true
-        } else {
-          return false
-        }
-      }),
-
-
-  data_documento: yup.string().required("Data do documento é obrigatório.").nullable(),
-  tipo_transacao: yup.string().required("Tipo da transação é obrigatório.").nullable(),
-
-  data_transacao: yup.string().required("Data do pagamento é obrigatório.").nullable(),
-  documento_transacao: yup.string().nullable(),
-
-  valor_total: yup.string().required("Valor do crédito é obrigatório.")
-    .test('test-valor', 'Valor deve ser maior que zero.',
-      function (value) {
-          return !(trataNumericos(value) <= 0);
-      })
-    .test('test-string', 'Valor do crédito é obrigatório.',
-      function (value) {
-          return !(typeof(value) == undefined)
-      }),
-
-  valor_recursos_proprios: yup.string().nullable(),
-  valor_total_dos_rateios:yup.string().nullable(),
-  valor_recusos_acoes:yup.string().nullable(),
-});
-
-export const periodoFechado = async (data, setReadOnlyBtnAcao, setShowPeriodoFechado, setReadOnlyCampos, onShowErroGeral) =>{
-  data = moment(data, "YYYY-MM-DD").format("YYYY-MM-DD");
-  try {
-    let periodo_fechado = await getPeriodoFechado(data);
-
-    if (!periodo_fechado.aceita_alteracoes){
-      setReadOnlyBtnAcao(true);
-      setShowPeriodoFechado(true);
-      setReadOnlyCampos(true);
-    }else {
-      setReadOnlyBtnAcao(false);
-      setShowPeriodoFechado(false);
-      setReadOnlyCampos(false);
-    }
-  }
-  catch (e) {
-    setReadOnlyBtnAcao(true);
-    setShowPeriodoFechado(true);
-    setReadOnlyCampos(true);
-    onShowErroGeral();
-    console.log("Erro ao buscar perído ", e)
-  }
-}
-
-export const periodoFechadoImposto = async (despesas_impostos, setReadOnlyBtnAcao, setShowPeriodoFechadoImposto, setReadOnlyCamposImposto, setDisableBtnAdicionarImposto, onShowErroGeral) =>{  
-  for(let despesa_imposto = 0; despesa_imposto <= despesas_impostos.length-1; despesa_imposto++){
-    if(despesas_impostos[despesa_imposto].data_transacao){
-      let data = moment(despesas_impostos[despesa_imposto].data_transacao, "YYYY-MM-DD").format("YYYY-MM-DD");
-
-      try{
-        let periodo_fechado = await getPeriodoFechado(data);
-        if (!periodo_fechado.aceita_alteracoes){
-          setReadOnlyBtnAcao(true);
-          setShowPeriodoFechadoImposto(true);
-          setReadOnlyCamposImposto(prevState => ({...prevState, [despesa_imposto]: true}));
-          setDisableBtnAdicionarImposto(true);
-        }
-        else{
-          setReadOnlyBtnAcao(false);
-          setShowPeriodoFechadoImposto(false);
-          setReadOnlyCamposImposto(prevState => ({...prevState, [despesa_imposto]: false}));
-          setDisableBtnAdicionarImposto(false);
-        }
-      }
-      catch (e){
-        setReadOnlyBtnAcao(true);
-        setShowPeriodoFechadoImposto(true);
-        setReadOnlyCamposImposto(prevState => ({...prevState, [despesa_imposto]: true}));
-        setDisableBtnAdicionarImposto(true);
-        onShowErroGeral();
-        console.log("Erro ao buscar perído ", e)
-      }
-    }
-  }
-}
-
-export const validaPayloadDespesas = (values, despesasTabelas=null, parametroLocation=null) => {
-  let exibe_documento_transacao
-  if (despesasTabelas){
-    if (typeof values.tipo_transacao === 'object' && values.tipo_transacao !== null){
-      exibe_documento_transacao = despesasTabelas.tipos_transacao.find(element => element.id === Number(values.tipo_transacao.id))
-    }else {
-      exibe_documento_transacao =  despesasTabelas.tipos_transacao.find(element => element.id === Number(values.tipo_transacao))
-    }
-    if(!values.tipo_transacao || !exibe_documento_transacao.tem_documento){
-      values.documento_transacao ="";
-    }
-  }
-
-  // Quando é Alteração
-  if (typeof values.associacao === "object"){
-    values.associacao = localStorage.getItem(ASSOCIACAO_UUID)
-  }
-
-  if (typeof values.tipo_documento === "object" && values.tipo_documento !== null){
-    values.tipo_documento = values.tipo_documento.id
-  }else {
-    if (values.tipo_documento !== "" && values.tipo_documento !== "0" && values.tipo_documento !== 0 && values.tipo_documento !== null) {
-      values.tipo_documento = convertToNumber(values.tipo_documento);
-    } else {
-      values.tipo_documento = null
-    }
-  }
-
-  if (typeof values.tipo_transacao === "object" && values.tipo_transacao !== null){
-    values.tipo_transacao = values.tipo_transacao.id
-  }else {
-    if (values.tipo_transacao !== "" && values.tipo_transacao !== "0" && values.tipo_transacao !== 0 && values.tipo_transacao !== null) {
-      values.tipo_transacao = convertToNumber(values.tipo_transacao);
-    } else {
-      values.tipo_transacao = null
-    }
-  }
-  values.valor_total = round(trataNumericos(values.valor_total),2);
-  values.valor_original = trataNumericos(values.valor_original);
-
-  values.valor_recursos_proprios = trataNumericos(values.valor_recursos_proprios);
-  values.valor_recusos_acoes = round((values.valor_recusos_acoes), 2)
-
-  if (values.data_documento !== "" && values.data_documento !== null){
-    values.data_documento = moment(values.data_documento).format("YYYY-MM-DD");
-  }else {
-    values.data_documento = null
-  }
-
-  if (values.data_transacao !== "" && values.data_transacao !== null){
-    values.data_transacao = moment(values.data_transacao).format("YYYY-MM-DD");
-  }else {
-    values.data_transacao = null
-  }
-
-  if (values.despesas_impostos){
-    // validacoes da despesa imposto
-    values.despesas_impostos.map((despesa_imposto) => {
-      if(despesa_imposto.data_transacao !== "" && despesa_imposto.data_transacao !== null){
-        despesa_imposto.data_transacao = moment(despesa_imposto.data_transacao).format("YYYY-MM-DD");
-      }
-      else{
-        despesa_imposto.data_transacao = null;
-      }
-
-      if(despesa_imposto.rateios.length >= 0){
-          despesa_imposto.rateios.map((rateio) => {
-              // o valor total e original da despesa imposto, devem ser o mesmo que o dos rateios
-              despesa_imposto.valor_total = trataNumericos(rateio.valor_rateio);
-              despesa_imposto.valor_original = trataNumericos(rateio.valor_original);
-
-              rateio.quantidade_itens_capital = convertToNumber(rateio.quantidade_itens_capital);
-              rateio.valor_item_capital = trataNumericos(rateio.valor_item_capital);
-              rateio.valor_rateio = round(trataNumericos(rateio.valor_rateio), 2);
-              rateio.valor_original = round(trataNumericos(rateio.valor_original), 2);
-          });
-
-          if(parametroLocation){
-            if(metodosAuxiliares.origemAnaliseLancamento(parametroLocation)){
-              metodosAuxiliares.mantemConciliacaoAtualImposto(despesa_imposto);
-            }
-          }
-      }
-    });
-  }
-
-  values.rateios.map((rateio) => {
-
-    if (typeof rateio.especificacao_material_servico === "object" && rateio.especificacao_material_servico !== null){
-      rateio.especificacao_material_servico = rateio.especificacao_material_servico.id
-    }else {
-      rateio.especificacao_material_servico = convertToNumber(rateio.especificacao_material_servico)
-    }
-
-    if (typeof rateio.conta_associacao === "object" && rateio.conta_associacao !== null){
-      rateio.conta_associacao = rateio.conta_associacao.uuid
-    }else {
-      if (rateio.conta_associacao === "0" || rateio.conta_associacao === "" || rateio.conta_associacao === 0){
-        rateio.conta_associacao = null
-      }
-    }
-
-    if (typeof rateio.acao_associacao === "object" && rateio.acao_associacao !== null){
-      rateio.acao_associacao = rateio.acao_associacao.uuid
-    }else {
-      if (rateio.acao_associacao === "0" || rateio.acao_associacao === "" || rateio.acao_associacao === 0) {
-        rateio.acao_associacao = null
-      }
-    }
-
-    if (rateio.tipo_custeio !== null ){
-
-      if (typeof rateio.tipo_custeio === "object" && rateio.tipo_custeio !== null){
-        rateio.tipo_custeio = rateio.tipo_custeio.id
-      }else {
-
-        if (rateio.tipo_custeio === "0" || rateio.tipo_custeio === 0 || rateio.tipo_custeio === ""){
-          rateio.tipo_custeio = null
-        }else {
-          rateio.tipo_custeio = convertToNumber(rateio.tipo_custeio)
-        }
-      }
-    }
-
-    if (typeof rateio.tag === "object" && rateio.tag !== null){
-      rateio.tag = rateio.tag.uuid
-    }else {
-      if ( rateio.tag === "" || rateio.escolha_tags === 'nao' ) {
-        rateio.tag = null
-      }
-    }
-
-    rateio.quantidade_itens_capital = convertToNumber(rateio.quantidade_itens_capital)
-    rateio.valor_item_capital = trataNumericos(rateio.valor_item_capital)
-    rateio.valor_rateio = round(trataNumericos(rateio.valor_rateio),2)
-    rateio.valor_original = round(trataNumericos(rateio.valor_original),2)
-
-    if (rateio.aplicacao_recurso === "0" || rateio.aplicacao_recurso === "" || rateio.aplicacao_recurso === 0){
-      rateio.aplicacao_recurso = null
-    }
-
-    if (rateio.especificacao_material_servico === "0" || rateio.especificacao_material_servico === 0 || rateio.especificacao_material_servico === ""){
-      rateio.especificacao_material_servico = null
-    }
-
-    if (rateio.aplicacao_recurso === "CAPITAL"){
-      //rateio.valor_rateio = round(rateio.quantidade_itens_capital * rateio.valor_item_capital, 2)
-    }
-
-  })
-
-  return values
-}
-
-export const exibeValorFormatadoPT_BR = (valor) => {
-  const isNegative = valor < 0;
-  const formattedValue = Math.abs(valor).toLocaleString('pt-BR', {
-    style: 'currency',
-    currency: 'BRL'
-  });
-
-  return isNegative ? `R$-${formattedValue.replace("R$", "").trim()}` : formattedValue;
-};
-
-export const exibeDataPT_BR = (data) => {
-  if (data === 'None'){
-    data = moment(new Date(), "YYYY-MM-DD").format("DD/MM/YYYY");
-  }else {
-    data =  moment(new Date(data), "YYYY-MM-DD").add(1, 'days').format("DD/MM/YYYY");
-  }
-  return data
-}
-
-export const exibeDateTimePT_BR = (data) => {
-  if (data === 'None'){
-    data = moment(new Date(), "YYYY-MM-DD").format("DD/MM/YYYY [às] HH:mm:ss");
-  }else {
-    data = moment(new Date(data), "YYYY-MM-DD").format("DD/MM/YYYY [às] HH:mm:ss");
-  }
-  return data
-}
-
-export const exibeDateTimePT_BR_Ata = (data) => {
-  if (data === 'None'){
-    data = moment(new Date(), "YYYY-MM-DD").format("DD/MM/YYYY [às] HH:mm");
-  }else {
-    data =  moment(new Date(data), "YYYY-MM-DD").format("DD/MM/YYYY [às] HH:mm");
-  }
-  return data
-}
-
-export const convertToNumber = (string)=>{
-  return Number(string)
-}
-
-export const round = (num, places) => {
-  return +(parseFloat(num).toFixed(places));
-}
-
-export const trataNumericos = (valor) => {
-  if (typeof valor === "string") {
-    // Extrai o sinal negativo, se existir.
-    const isNegative = valor.startsWith('-');
-
-    // Remove todos os caracteres não numéricos, exceto a vírgula.
-    const cleaned = valor.replace(/[^\d,]/g, "");
-
-    // Converte a vírgula em um ponto para torná-la uma string numérica válida.
-    const converted = cleaned.replace(/,/g, ".");
-
-    // Converte a string numérica em um número.
-    const numberValue = Number(converted);
-
-    // Aplica de volta o sinal negativo, se necessário.
-    return isNegative ? -numberValue : numberValue;
-  } else {
-    return valor;
-  }
-}
-
-export const calculaValorRateio = (valor1, valor2) => {
-
-  let valor1Tratado = trataNumericos(valor1)
-  let valor2Tratado = trataNumericos(valor2)
-
-  let valor_total = valor1Tratado * valor2Tratado;
-
-  return valor_total;
-};
-export const calculaValorRecursoAcoes = (values) => {
-  let valor_totalTratado = trataNumericos(values.valor_total);
-  let valor_recursos_propriosTratado = trataNumericos(values.valor_recursos_proprios);
-  return round(valor_totalTratado - valor_recursos_propriosTratado, 2);
-};
-
-export const calculaValorOriginal = (values) => {
-  let valor_total_ratado = trataNumericos(values.valor_original);
-  let valor_recursos_proprios_tratado = trataNumericos(values.valor_recursos_proprios);
-  let valor_total = round(valor_total_ratado - valor_recursos_proprios_tratado, 2);
-  return valor_total;
-};
-
-const PESOS_DV_CNPJ = [6, 5, 4, 3, 2, 9, 8, 7, 6, 5, 4, 3, 2];
-const CNPJ_ZERADO = "00000000000000";
-const CNPJ_MASK = [
-  /[A-Za-z0-9]/, /[A-Za-z0-9]/, '.',
-  /[A-Za-z0-9]/, /[A-Za-z0-9]/, /[A-Za-z0-9]/, '.',
-  /[A-Za-z0-9]/, /[A-Za-z0-9]/, /[A-Za-z0-9]/, '/',
-  /[A-Za-z0-9]/, /[A-Za-z0-9]/, /[A-Za-z0-9]/, /[A-Za-z0-9]/, '-',
-  /\d/, /\d/
-];
-
-const removeMascaraCnpj = (valor) => {
-  return valor.replace(/[.\-/]/g, "").toUpperCase();
-};
-
-const charValueCnpj = (caractere) => {
-  return caractere.toUpperCase().charCodeAt(0) - 48;
-};
-
-const calculaDvCnpj = (base12) => {
-  let somatorioDv1 = 0;
-  let somatorioDv2 = 0;
-
-  for (let i = 0; i < 12; i++) {
-    const valorCaractere = charValueCnpj(base12[i]);
-    somatorioDv1 += valorCaractere * PESOS_DV_CNPJ[i + 1];
-    somatorioDv2 += valorCaractere * PESOS_DV_CNPJ[i];
-  }
-
-  const dv1 = somatorioDv1 % 11 < 2 ? 0 : 11 - (somatorioDv1 % 11);
-  somatorioDv2 += dv1 * PESOS_DV_CNPJ[12];
-  const dv2 = somatorioDv2 % 11 < 2 ? 0 : 11 - (somatorioDv2 % 11);
-
-  return `${dv1}${dv2}`;
-};
-
-export const cpfMaskContitional = (value) => {
-  const cpfCnpj = value.replace(/[^A-Za-z0-9]+/g, "").toUpperCase();
-
-  if (cpfCnpj.length > 11 || /[A-Za-z]/.test(cpfCnpj)) {
-    return CNPJ_MASK;
-  }
-
-  return [/\d/, /\d/, /\d/, '.', /\d/, /\d/, /\d/, '.', /\d/, /\d/, /\d/, '-', /\d/, /\d/];
-};
-
-export const processoIncorporacaoMask = (value) => {
-  // 0000.0000/0000000-0
-  let processo = value.replace(/[^\d]+/g, "");
-
-  let mask = [/\d/, /\d/, /\d/, /\d/, '.', /\d/, /\d/, /\d/, /\d/, '/', /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, '-', /\d/]
-
-  return mask
-}
-
-const limpaCpf = (valor) => {
-  return valor ? valor.replace(/[^0-9]/g, '') : '';
-};
-
-const limpaCpfCnpj = (valor) => {
-  return valor ? valor.replace(/[^A-Za-z0-9]/g, '').toUpperCase() : '';
-};
-
-const possuiSequenciaInvalida = (valor, permitirCnpjZerado = false) => {
-  const sequenciasInvalidas = [
-    "00000000000",
-    "11111111111",
-    "11111111111111",
-    "22222222222",
-    "22222222222222",
-    "33333333333",
-    "33333333333333",
-    "44444444444",
-    "44444444444444",
-    "55555555555",
-    "55555555555555",
-    "66666666666",
-    "66666666666666",
-    "77777777777",
-    "77777777777777",
-    "88888888888",
-    "88888888888888",
-    "99999999999",
-    "99999999999999",
-  ];
-
-  if (!permitirCnpjZerado) {
-    sequenciasInvalidas.push("00000000000000");
-  }
-
-  return sequenciasInvalidas.includes(valor);
-};
-
-function verifica_cpf_cnpj(valor) {
-  const valorCpf = limpaCpf(valor);
-  const valorCpfCnpj = limpaCpfCnpj(valor);
-
-  if (valorCpf.length === 11 && /^\d+$/.test(valorCpf)) {
-    return 'CPF';
-  }
-
-  if (valorCpfCnpj.length === 14) {
-    return 'CNPJ';
-  }
-
-  return false;
-}
-
-export function valida_cpf_cnpj_permitindo_cnpj_zerado(valor) {
-  const tipoDocumento = verifica_cpf_cnpj(valor);
-
-  if (!tipoDocumento) {
-    return false;
-  }
-
-  if (tipoDocumento === 'CPF') {
-    const cpf = limpaCpf(valor);
-    if (possuiSequenciaInvalida(cpf, true)) {
-      return false;
-    }
-    return valida_cpf(cpf);
-  }
-
-  const cnpj = limpaCpfCnpj(valor);
-  if (cnpj === CNPJ_ZERADO) {
-    return true;
-  }
-  if (possuiSequenciaInvalida(cnpj, true)) {
-    return false;
-  }
-  return valida_cnpj(cnpj);
-}
-
-export function valida_cpf_cnpj(valor) {
-  const tipoDocumento = verifica_cpf_cnpj(valor);
-
-  if (!tipoDocumento) {
-    return false;
-  }
-
-  if (tipoDocumento === 'CPF') {
-    const cpf = limpaCpf(valor);
-    if (possuiSequenciaInvalida(cpf)) {
-      return false;
-    }
-    return valida_cpf(cpf);
-  }
-
-  const cnpj = limpaCpfCnpj(valor);
-  if (possuiSequenciaInvalida(cnpj)) {
-    return false;
-  }
-  return valida_cnpj(cnpj);
-}
-
-function calc_digitos_posicoes( digitos, posicoes = 10, soma_digitos = 0 ) {
-
-  digitos = digitos.toString();
-
-  for ( let i = 0; i < digitos.length; i++  ) {
-    soma_digitos = soma_digitos + ( digitos[i] * posicoes );
-    posicoes--;
-    if ( posicoes < 2 ) {
-      posicoes = 9;
-    }
-  }
-
-  soma_digitos = soma_digitos % 11;
-
-  if ( soma_digitos < 2 ) {
-    soma_digitos = 0;
-  } else {
-    soma_digitos = 11 - soma_digitos;
-  }
-
-  let cpf = digitos + soma_digitos;
-
-  return cpf;
-
-} // calc_digitos_posicoes
-
-function valida_cpf( valor ) {
-
-  valor = valor.toString();
-
-  valor = valor.replace(/[^0-9]/g, '');
-
-  let digitos = valor.substr(0, 9);
-
-  let novo_cpf = calc_digitos_posicoes( digitos );
-
-  novo_cpf = calc_digitos_posicoes( novo_cpf, 11 );
-
-  if ( novo_cpf === valor ) {
-    return true;
-  } else {
-    return false;
-  }
-
-} // valida_cpf
-
-export function valida_cnpj(valor) {
-  if (!valor) {
-    return false;
-  }
-
-  const cnpj = removeMascaraCnpj(valor.toString());
-
-  if (cnpj === CNPJ_ZERADO) {
-    return false;
-  }
-
-  if (!/^[A-Z0-9]{12}\d{2}$/.test(cnpj)) {
-    return false;
-  }
-
-  return cnpj.slice(-2) === calculaDvCnpj(cnpj.slice(0, 12));
-}
-
-export const getTextoStatusPeriodo = (statusId) => {
-  if (statusId === 'EM_ANDAMENTO') {
-    status = 'O período está em andamento, os dados apresentados estão em atualização sendo cadastrados.'
-  } else if (statusId === 'PENDENTE') {
-    status = 'O período está pendente'
-  } else if (statusId === 'CONCILIADO') {
-    status = 'O período foi conferido e fechado pela Associação de Pais e Mestres'
-  } else if (statusId === 'APROVADO') {
-    status = 'O período está fechado e foi aprovado pela Diretoria Regional de Educação'
-  } else if (statusId === 'REJEITADO') {
-    status =
-        'O período está fechado e foi rejeitado pela Diretoria Regional de Educação'
-  } else {
-    status = 'O período está com status indefinido'
-  }
-  return status
-};
-
-export const getCorStatusPeriodo = (statusId) => {
-  let cor = ''
-  if (statusId === 'EM_ANDAMENTO' || (statusId && statusId.prestacao_contas_status && statusId.prestacao_contas_status.texto_status === "Período em andamento. ")) {
-    cor = 'amarelo'
-  } else if (statusId === 'PENDENTE') {
-    cor = 'vermelho'
-  } else if (statusId === 'CONCILIADO') {
-    cor = 'azul'
-  } else if (statusId === 'APROVADO') {
-    cor = 'verde'
-  } else if (statusId === 'REJEITADO') {
-    cor = 'vermelho'
-  } else {
-    cor = 'vermelho'
-  }
-  return cor
-};
-
-
-export const slugify = (string) =>{
-  return string
-  .toString()
-  .trim()
-  .toLowerCase()
-  .replace(/[àÀáÁâÂãäÄÅåª]+/g, 'a')       // Special Characters #1
-  .replace(/[èÈéÉêÊëË]+/g, 'e')       	// Special Characters #2
-  .replace(/[ìÌíÍîÎïÏ]+/g, 'i')       	// Special Characters #3
-  .replace(/[òÒóÓôÔõÕöÖº]+/g, 'o')       	// Special Characters #4
-  .replace(/[ùÙúÚûÛüÜ]+/g, 'u')       	// Special Characters #5
-  .replace(/[ýÝÿŸ]+/g, 'y')       		// Special Characters #6
-  .replace(/[ñÑ]+/g, 'n')       			// Special Characters #7
-  .replace(/[çÇ]+/g, 'c')       			// Special Characters #8
-  .replace(/[ß]+/g, 'ss')       			// Special Characters #9
-  .replace(/[Ææ]+/g, 'ae')       			// Special Characters #10
-  .replace(/[Øøœ]+/g, 'oe')       		// Special Characters #11
-  .replace(/[%]+/g, 'pct')       			// Special Characters #12
-  .replace(/\s+/g, '-')           		// Replace spaces with -
-  .replace(/[^\w\-]+/g, '')       		// Remove all non-word chars
-  .replace(/\-\-+/g, '-')         		// Replace multiple - with single -
-  .replace(/^-+/, '')             		// Trim - from start of text
-  .replace(/-+$/, '');            		// Trim - from end of text
-};
-
-export const gerarUuid = () => {
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-    var r = Math.random()*16|0, v = c == 'x' ? r : (r&0x3|0x8);
-    return v.toString(16)
-  });
-};
-
-export const comparaObjetos = (objetoA, objetoB) =>{
-
-  if (objetoA === null  || objetoB === null ){
-    // null é um 'object', mas não é possível se obter keys de um null. Tratamento precisa ser distinto.
-    return objetoA === objetoB
-  }
-
-  let aChaves = Object.keys(objetoA),
-      bChaves = Object.keys(objetoB);
-
-  if (aChaves.length !== bChaves.length) {
-    return false;
-  }
-
-  let saoDiferentes = aChaves.some((chave) => {
-    if (typeof objetoA[chave] == "object" && typeof objetoB[chave] == "object" ){
-      return !comparaObjetos(objetoA[chave] , objetoB[chave] )
-    }
-    else {
-      return objetoA[chave] !== objetoB[chave];
-    }
-  });
-
-  return !saoDiferentes;
-}
-
-export const valida_cpf_exportado = ( valor ) => {
-
-  valor = valor.toString();
-
-  valor = valor.replace(/[^0-9]/g, '');
-
-  let digitos = valor.substr(0, 9);
-
-  let novo_cpf = calc_digitos_posicoes( digitos );
-
-  novo_cpf = calc_digitos_posicoes( novo_cpf, 11 );
-
-  if ( novo_cpf === valor ) {
-    return true;
-  } else {
-    return false;
-  }
-
-} // valida_cpf
-
-export const apenasNumero = (valor) => {
-	const re = /^[0-9\b]+$/;
-	
-	if (valor === '' || re.test(valor)) {
-		return true;
-		
-	}
-	return false;
-}
-
-export const formataNomeDRE = (nome_dre) => {
-  if(nome_dre !== ""){
-    nome_dre = nome_dre.toUpperCase();
-    
-    if(nome_dre.indexOf("DIRETORIA REGIONAL DE EDUCACAO") !== -1){
-      nome_dre = nome_dre.replace("DIRETORIA REGIONAL DE EDUCACAO", "");
-      nome_dre = nome_dre.trim();
-      
-      return nome_dre;
-    }
-  }
-  
-
-  return nome_dre;
-}
